requirements.txt
matplotlib==3.2.2
nltk==3.5
pandas==1.0.5
Pillow==8.4.0
plotly==5.1.0
seaborn==0.10.1
streamlit==1.0.0
wordcloud==1.8.1
